+++ mainlib/ant/antsmarttrainerchannelhandler.cpp
@@ -62,6 +62,13 @@ private:
     State _state;
 };
 
+static const auto STATE_STRINGS = std::map<GeneralFitnessEquipmentMessage::State,QString>(
+    {{GeneralFitnessEquipmentMessage::State::ASLEEP, QString("ASLEEP")},
+     {GeneralFitnessEquipmentMessage::State::READY, "READY"},
+     {GeneralFitnessEquipmentMessage::State::IN_USE, "IN_USE"},
+     {GeneralFitnessEquipmentMessage::State::FINISHED, "FINISHED"}});
+
+
 /**
  * GeneralFitnessEquipmentMessage is an ANT+ Broadcast message.
  *
@@ -216,6 +223,20 @@ private:
     qreal _userWeight;
 };
 
+/**
+ * Capabilities Message, containing all capabilities of the FE-C trainer.
+ */
+class CapabilitiesMessage: public BroadCastMessage
+{
+public:
+    explicit CapabilitiesMessage(const AntMessage2& antMessage);
+
+    const bool simulationModeSupported;
+    const int maxResistanceNewtons;
+private:
+    bool parseSimulationModeSupportedCapability();
+};
+
 AntSmartTrainerChannelHandler::AntSmartTrainerChannelHandler(int channelNumber, QObject *parent) :
     AntChannelHandler(channelNumber, AntSensorType::SMART_TRAINER, AntSportPeriod::SMART_TRAINER, parent),
     _configurationCheckerTimer(new QTimer(this))
@@ -238,6 +259,9 @@ void AntSmartTrainerChannelHandler::handleBroadCastMessage(const BroadCastMessag
 //    case DataPage::TRACK_RESISTANCE:
 //        handleTrackResistanceMessage(TrackResistanceMessage(message.antMessage()));
 //        break;
+    case DataPage::FE_CAPABILITIES:
+        handleCapabilitiesMessage(CapabilitiesMessage(message.antMessage()));
+        break;
     case DataPage::COMMAND_STATUS:
         handleCommandStatusReplyMessage(CommandStatusReplyMessage(message.antMessage()));
         break;
@@ -254,20 +278,41 @@ void AntSmartTrainerChannelHandler::handleBroadCastMessage(const BroadCastMessag
 
 void AntSmartTrainerChannelHandler::checkConfiguration()
 {
-    queueAcknowledgedMessage(createRequestMessage(DataPage::COMMAND_STATUS));
+    // FIXME (IB): we won't send anything right now. I don't really get
+    // how this should work.
+    //    queueAcknowledgedMessage(createRequestMessage(DataPage::COMMAND_STATUS));
+}
+
+void AntSmartTrainerChannelHandler::setState(AntSmartTrainerChannelHandler::State state)
+{
+    _state = state;
+    qDebug() << channelIdString() << "Going to state" << static_cast<int>(state);
+}
+
+void AntSmartTrainerChannelHandler::handleCapabilitiesMessage(const CapabilitiesMessage &message)
+{
+    if (_state == State::CAPABILITIES_REQUESTED) {
+        qDebug() << "max trainer resistance = " << QString("%1 N").arg(message.maxResistanceNewtons);
+        qDebug() << "simulation mode supported" << message.simulationModeSupported;
+        setState(State::INITIALIZED);
+        sendWindResistanceMessage();
+        sendTrackResistanceMessage();
+        sendUserConfigurationMessage();
+        _configurationCheckerTimer->start();
+    }
 }
 
 void AntSmartTrainerChannelHandler::setWeight(const qreal userWeightInKilograms, const qreal bikeWeightInKilograms)
 {
     _userWeight = userWeightInKilograms;
     _bikeWeight = bikeWeightInKilograms;
-    queueAcknowledgedMessage(createUserConfigurationMessage());
+    sendUserConfigurationMessage();
 }
 
 void AntSmartTrainerChannelHandler::setSlope(const qreal slopeInPercent)
 {
     _slope = slopeInPercent;
-    queueAcknowledgedMessage(createTrackResistanceMessage());
+    sendTrackResistanceMessage();
 }
 
 void AntSmartTrainerChannelHandler::close()
@@ -278,14 +323,12 @@ void AntSmartTrainerChannelHandler::close()
 
 void AntSmartTrainerChannelHandler::channelOpened()
 {
-    queueAcknowledgedMessage(createWindResistenceMessage());
-    queueAcknowledgedMessage(createTrackResistanceMessage());
-    _configurationCheckerTimer->start();
+    requestCapabilities();
 }
 
 void AntSmartTrainerChannelHandler::handleGeneralFitnessEquipmentMessage(const GeneralFitnessEquipmentMessage &message)
 {
-    qDebug() << "state = " << static_cast<int>(message.state());
+    qDebug() << channelIdString() << "channel state = " << STATE_STRINGS.at(message.state());
 }
 
 void AntSmartTrainerChannelHandler::handleSpecificTrainerDataMessage(const SpecificTrainerDataMessage &message)
@@ -294,33 +337,60 @@ void AntSmartTrainerChannelHandler::handleSpecificTrainerDataMessage(const Speci
     emit sensorValue(SensorValueType::CADENCE_RPM, AntSensorType::SMART_TRAINER, QVariant::fromValue(message.cadence()));
 
     if (message.userConfigurationNeeded()) {
-        queueAcknowledgedMessage(createUserConfigurationMessage());
+        qDebug() << channelIdString() << "User configuration needed. Queueing user configuration message";
+        sendUserConfigurationMessage();
     }
 }
 
 void AntSmartTrainerChannelHandler::handleCommandStatusReplyMessage(const CommandStatusReplyMessage &message)
 {
-    qDebug("Last received command id = %d, Sequence # = %d, Status = %d", message.received(), message.sequenceNr(), static_cast<quint8>(message.status()));
-    queueAcknowledgedMessage(createRequestMessage(DataPage::WIND_RESISTANCE));
-    queueAcknowledgedMessage(createRequestMessage(DataPage::TRACK_RESISTANCE));
+    qDebug("%s Last received command id = %d, Sequence # = %d, Status = %d", qPrintable(channelIdString()), message.received(), message.sequenceNr(), static_cast<quint8>(message.status()));
 }
 
 void AntSmartTrainerChannelHandler::handleTrackResistanceMessage(const TrackResistanceMessage &message)
 {
-    qDebug() << "Track resistance, slope = " << message.slope();
+    qDebug() << channelIdString() << "Track resistance, slope = " << message.slope();
 }
 
 void AntSmartTrainerChannelHandler::handleWindResistanceMessage(const WindResistanceMessage &message)
 {
-    qDebug() << "wind resistance" << message.windResistanceCoeffient() << message.windSpeed() << message.draftingFactor();
+    qDebug() << channelIdString() << "wind resistance" << message.windResistanceCoeffient() << message.windSpeed() << message.draftingFactor();
 }
 
 void AntSmartTrainerChannelHandler::handleUserConfigurationMessage(const UserConfigurationMessage &message)
 {
     if (qFuzzyCompare(message.userWeight(), _userWeight)) {
-        qDebug() << "User weight correctly configured";
+        qDebug() << channelIdString() << "User weight correctly configured";
     } else {
-        qWarning("User weight not correctly configured on device!");
+        qWarning("%s User weight not correctly configured on device!", qPrintable(channelIdString()));
+    }
+}
+
+void AntSmartTrainerChannelHandler::requestCapabilities()
+{
+    AntMessage2 requestCapabiltiesMessage = createRequestMessage(DataPage::FE_CAPABILITIES);
+    queueAcknowledgedMessage(requestCapabiltiesMessage);
+    setState(State::CAPABILITIES_REQUESTED);
+}
+
+void AntSmartTrainerChannelHandler::sendWindResistanceMessage()
+{
+    if (_state == State::INITIALIZED) {
+        queueAcknowledgedMessage(createWindResistenceMessage());
+    }
+}
+
+void AntSmartTrainerChannelHandler::sendTrackResistanceMessage()
+{
+    if (_state == State::INITIALIZED) {
+        queueAcknowledgedMessage(createTrackResistanceMessage());
+    }
+}
+
+void AntSmartTrainerChannelHandler::sendUserConfigurationMessage()
+{
+    if (_state == State::INITIALIZED) {
+        queueAcknowledgedMessage(createUserConfigurationMessage());
     }
 }
 
@@ -415,7 +485,8 @@ AntMessage2 AntSmartTrainerChannelHandler::createUserConfigurationMessage()
     quint8 bikeWeightMsb = ((bikeWeight >> 4) & 0xFF);
     content += bikeWeightMsb;
 
-    content += 0xFF; // default wheel size (0.7m diameter, 700C, normal racing wheels.
+    quint8 wheelDiameter = 70u;
+    content += wheelDiameter; // default wheel size (0.7m diameter, 700C, normal racing wheels.
     quint8 invalidGearRatio = 0x0;
     content += invalidGearRatio; // invalid gear ratio. Not needed.
 
@@ -431,7 +502,7 @@ AntMessage2 AntSmartTrainerChannelHandler::createRequestMessage(DataPage dataPag
     content += 0xFF; // reserved
     content += 0xFF; // no descriptor
     content += 0xFF; // no descriptor
-    content += 0x01; // transmit the page once.
+    content += 0x08; // transmit the page eight times.
     content += static_cast<quint8>(dataPage);
     content += static_cast<quint8>(0x1);
 
@@ -516,10 +587,10 @@ bool SpecificTrainerDataMessage::userConfigurationNeeded() const
 CommandStatusReplyMessage::CommandStatusReplyMessage(const AntMessage2 &antMessage):
     BroadCastMessage(antMessage)
 {
-    _dataPage = static_cast<AntSmartTrainerChannelHandler::DataPage>(antMessage.contentByte(1));
-    _lastCommand = antMessage.contentByte(2);
-    _sequenceNr = antMessage.contentByte(3);
-    _status = static_cast<Status>(antMessage.contentByte(4));
+    _dataPage = static_cast<AntSmartTrainerChannelHandler::DataPage>(contentByte(0));
+    _lastCommand = contentByte(1);
+    _sequenceNr = contentByte(2);
+    _status = static_cast<Status>(contentByte(3));
 }
 
 UserConfigurationMessage::UserConfigurationMessage(const AntMessage2 &antMessage)
@@ -541,6 +612,20 @@ TrackResistanceMessage::TrackResistanceMessage(const AntMessage2 &antMessage):
 {
     quint16 slopeAsInt = contentShort(5);
     _slope = (slopeAsInt / 100.0) - 200.0;
+}
+
+CapabilitiesMessage::CapabilitiesMessage(const AntMessage2 &antMessage):
+    BroadCastMessage(antMessage),
+    simulationModeSupported(parseSimulationModeSupportedCapability()),
+    maxResistanceNewtons(static_cast<int>(antMessage.contentShort(5)))
+{
+    // empty
+}
+
+bool CapabilitiesMessage::parseSimulationModeSupportedCapability()
+{
+    quint8 capabilitiesByte = contentByte(7);
+    return capabilitiesByte & 4;
 }
 
 }