+++ mainlib/config/bigringsettings.h
@@ -33,7 +33,12 @@ public:
     BigRingSettings();
 
     QString videoFolder() const;
-    void setVideoFolder(const QString folder);
+    void setVideoFolder(const QString &folder);
+
+    /** Get the absolute path to folder where tcx ride files are saved */
+    QString tcxFolder() const;
+    /** Set the absolute path to the folder where tcx ride files are saved */
+    void setTcxFolder(const QString &folder);
 
     qreal userWeight() const;
     void setUserWeight(const qreal userWeight);
@@ -43,6 +48,9 @@ public:
 
     int powerAveragingForDisplayMilliseconds() const;
     void setPowerAveragingForDisplayMilliseconds(const int averagingMilliseconds);
+
+    /** Get the unique id for this installation */
+    QString clientId();
 private:
     QSettings _settings;
 };