+++ mainlib/maingui/mainwindow.h
@@ -30,6 +30,7 @@
 #include "importer/reallifevideoimporter.h"
 #include "util/util.h"
 
+class AnalyticsSender;
 class Cyclist;
 class VideoListView;
 class NewVideoWidget;
@@ -56,12 +57,14 @@ protected:
     virtual void closeEvent(QCloseEvent *);
     virtual void showEvent(QShowEvent *);
 private slots:
+    void initialize();
     void loadVideos();
     void importFinished(RealLifeVideoList rlvs);
     void removeDisplayMessage();
+    /** Show that a new version is available */
+    void newVersionAvailable(bool newVersion, const QString &version);
 private:
     void setupMenuBar();
-    void step();
     void startRun(RealLifeVideo rlv, int courseNr);
 
     indoorcycling::AntCentralDispatch* const _antCentralDispatch;
@@ -74,7 +77,10 @@ private:
 
     VideoListView* const _listView;
     QScopedPointer<NewVideoWidget> _videoWidget;
+    bool _guiFullScreen;
     QRect _savedGeometry;
+
+    AnalyticsSender *_analyticsSender;
 };
 
 #endif // MAINWINDOW_H