+++ mainlib/maingui/settingsdialog.cpp
@@ -54,15 +54,15 @@ SettingsDialog::SettingsDialog(indoorcycling::AntCentralDispatch* antCentralDisp
                                QWidget *parent) :
     QDialog(parent),
     _ui(new Ui::SettingsDialog),
+    _quantityPrinter(new QuantityPrinter(this)),
+    _unitConverter(new UnitConverter(this)),
     _antCentralDispatch(antCentralDispatch),
     _videoLoadFunction(videoLoadFunction)
 {
     _ui->setupUi(this);
     QSettings settings;
     _ui->unitChooser->setCurrentText(settings.value("units").toString());
-    BigRingSettings bigRingSettings;
-    _ui->userWeightSpinBox->setValue(bigRingSettings.userWeight());
-    _ui->bikeWeightSpinBox->setValue(bigRingSettings.bikeWeight());
+
 
     reset();
 }
@@ -76,6 +76,7 @@ void SettingsDialog::on_unitChooser_currentTextChanged(const QString &choice)
 {
     QSettings settings;
     settings.setValue("units", choice);
+    reset();
 }
 
 void SettingsDialog::on_pushButton_clicked()
@@ -169,10 +170,30 @@ void SettingsDialog::fillVideoFolderList()
     _ui->videoFolderLineEdit->setText(_settings.videoFolder());
 }
 
+void SettingsDialog::fillWeights()
+{
+    _ui->userWeightSpinBox->blockSignals(true);
+    _ui->bikeWeightSpinBox->blockSignals(true);
+
+    const qreal userWeight = _unitConverter->convertWeightToSystemUnit(_settings.userWeight());
+    _ui->userWeightSpinBox->setValue(userWeight);
+    const qreal bikeWeight = _unitConverter->convertWeightToSystemUnit(_settings.bikeWeight());
+    _ui->bikeWeightSpinBox->setValue(bikeWeight);
+
+    const QString unitString = QString(" %1").arg(_quantityPrinter->unitForWeight());
+
+    _ui->userWeightSpinBox->setSuffix(unitString);
+    _ui->bikeWeightSpinBox->setSuffix(unitString);
+
+    _ui->userWeightSpinBox->blockSignals(false);
+    _ui->bikeWeightSpinBox->blockSignals(false);
+}
+
 void SettingsDialog::fillPowerAveragingComboBox()
 {
     _ui->powerAveragingCombobox->blockSignals(true);
 
+    _ui->powerAveragingCombobox->clear();
     _ui->powerAveragingCombobox->addItem(tr("No Averaging"), 0);
     _ui->powerAveragingCombobox->addItem(tr("1 Second"), 1000);
     _ui->powerAveragingCombobox->addItem(tr("3 Seconds"), 3000);
@@ -199,6 +220,24 @@ void SettingsDialog::saveVideoFolder(const QString& folder)
     _videoLoadFunction();
 }
 
+bool SettingsDialog::ableToWriteInFolder(const QString &folder)
+{
+    // try to write and remove a file in the directory. If that does not work, permissions might be off.
+    QDir tcxDirectory(folder);
+    const QString testFilePath = tcxDirectory.absoluteFilePath("testfile");
+    QFile testFile(testFilePath);
+    if (testFile.open(QFile::ReadWrite)) {
+        // remove the test file again.
+        testFile.close();
+        testFile.remove();
+        return true;
+    } else {
+        QMessageBox::warning(this, tr("Unable to write in folder"),
+                             tr("Unable to write in folder %1, please choose another folder or adjust permissions.").arg(folder));
+        return false;
+    }
+}
+
 void SettingsDialog::on_antConfigurationChooser_currentIndexChanged(
         const QString &selectedConfiguration)
 {
@@ -214,6 +253,9 @@ void SettingsDialog::reset()
     fillSimulationSettingLabel();
     fillVideoFolderList();
     fillPowerAveragingComboBox();
+    fillWeights();
+
+    _ui->tcxSaveLocationTextEdit->setText(_settings.tcxFolder());
 }
 
 void SettingsDialog::on_deleteConfigurationButton_clicked()
@@ -247,15 +289,37 @@ void SettingsDialog::on_powerAveragingCombobox_currentIndexChanged(int index)
     QVariant data = _ui->powerAveragingCombobox->itemData(index);
     // values are 1, 3 and 10 seconds.
     _settings.setPowerAveragingForDisplayMilliseconds(data.toInt());
-
 }
 
 void SettingsDialog::on_userWeightSpinBox_valueChanged(double userWeight)
 {
-    BigRingSettings().setUserWeight(userWeight);
+    const double weightInKilograms = _unitConverter->convertWeightFromSystemUnit(userWeight);
+    BigRingSettings().setUserWeight(weightInKilograms);
 }
 
 void SettingsDialog::on_bikeWeightSpinBox_valueChanged(double bikeWeight)
 {
-    BigRingSettings().setBikeWeight(bikeWeight);
+    const double weightInKilograms = _unitConverter->convertWeightFromSystemUnit(bikeWeight);
+    BigRingSettings().setBikeWeight(weightInKilograms);
+}
+
+void SettingsDialog::on_changeTcxFolderButton_clicked()
+{
+    const QStringList homeDirectories = QStandardPaths::standardLocations(QStandardPaths::HomeLocation);
+
+    const QString startDirectory = (homeDirectories.isEmpty()) ? QString() : homeDirectories[0];
+    const QString dir = QFileDialog::getExistingDirectory(this, tr("Choose Ride File Directory"),
+                                                    startDirectory);
+    if (dir.isEmpty()) {
+        // file dialog was cancelled.
+        return;
+    }
+
+    if (ableToWriteInFolder(dir)) {
+        _settings.setTcxFolder(dir);
+        _ui->tcxSaveLocationTextEdit->setText(dir);
+    } else {
+        // open the file dialog again. We'll put this on the event loop to avoid blocking.
+        QTimer::singleShot(0, this, &SettingsDialog::on_changeTcxFolderButton_clicked);
+    }
 }